pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Final Project for the Simulation and Performance Evaluation course"
license = "MIT"
authors = ["Manuela Corte Pause <manuela.cortepause@studenti.unitn.it>",
"Simone Marrocco <simone.marrocco@studenti.unitn.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
statsmodels = "^0.14.1"
seaborn = {extras = ["stats"], version = "^0.13.1"}
graphviz = "^0.20.1"
tqdm = "^4.66.4"
pyqt6 = "^6.7.0"
ipykernel = "^6.29.4"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pre-commit = "^3.6.0"
types-seaborn = "^0.13.1.20240115"
pandas-stubs = "^2.1.4.231227"
pyright = "^1.1.349"

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules",
    "**/__pycache__",
]

typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnusedImport = "information"
reportUnusedVariable = "information"
reportUnnecessaryComparison = "information"

pythonVersion = "3.12"
pythonPlatform = "Linux"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
